Improve the path description of a file
So far, these docs don't use a preceeding forward slash (i.e /) when describing paths relative to the root of a project. This commit fixes a particular path that mistakenly does, because its inconsistency with other paths might signal to a Linux or Mac user that that path is with respect to the root of his/her filesystem.