Use sudo in Linux installation docs
The current Linux installation docs do not differentiate between
commands that need to run as root versus those that don't.  The issue
here is that if everything is run as root then reaction will be run as
root, which shouldn't be necessary. This should hopefully clarify what
to run as root versus the commands that can be run as an unprivileged
user.