Remove sudo call on installation of reaction-cli
The installation of reaction-cli needs to be done globally, and for
some users this will require the use of sudo.  However, this is
dependent on how Node.js has been installed so it may be better for
the user to try without sudo first in the event that it works without
needing to sudo.