General doc cleanup
The link /admin/email.md doesn't exist, so we update links to point to
/admin/dashboard/email-admin.md.

The link configuration.md doesn't exist, so we point to
/developer/configuration.md.

The link /developer/tutorial/customization.md was renamed to
/developer/tutorial/introduction.md in a6769b8, so we update the
reference here.

The link /developer/tutorial/creating-a-plugin.md doesn't exist, so we
point to /developer/tutorial/plugin-creating-2.md.