fix: add stance on Windows development support
Developing Reaction Commerce in a Windows environment has been difficult. Because the development environment is containerized, in theory development in Windows should be straightforward. In practice, there has been difficulty developing in Windows. We’d like to be able to support development in Windows and plan to revisit the policy soon. 

Signed-off-by: Spencer Norman <spencer@reactioncommerce.com>